feat!: `start`, `stop`, `destroy` ready constraints is loosened
When ready methods are called, they may run in between the lock
acquisition and the status change.

In the case of synchronous methods, they should just proceed to execute
without resulting in an error. This is because they are capable of
completely finishing by the time the status change occurs, so it is safe
to run them, even without the `allowedStatuses` exclusions. This also
applies to property getters and setters.

In the case of asynchronous methods, they need to wait for the status to
be changed before they can proceed to check for `allowedStatuses`
exclusions. This is because their execution may overlap with the
`start`, `stop`, and `destroy` calls.

Originally there would be a bug where if an asynchronous method were to
be executed in between lock acquisition and the status change, the
method would fail with a ready exception even if with `allowedStatuses`
exclusions. This no longer happens.